pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "glassbox-rag"
version = "0.1.0"
description = "An open-source, high-transparency modular RAG framework for AI/ML applications"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "GlassBox Team", email = "team@glassbox.ai"}
]
keywords = ["rag", "retrieval", "generation", "onnx", "ollama", "vector-store", "multimodal"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pyyaml>=6.0",
    "numpy>=1.24.0",
    "requests>=2.31.0",
    "python-multipart>=0.0.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
]

embeddings = [
    "onnxruntime>=1.16.0",
    "ollama>=0.0.50",
    "openai>=1.3.0",
]

vector-stores = [
    "qdrant-client>=2.4.0",
    "chromadb>=0.4.0",
    "psycopg2-binary>=2.9.0",
]

multimodal = [
    "pillow>=10.0.0",
    "pypdf>=3.17.0",
    "python-pptx>=0.6.21",
]

[project.scripts]
glassbox-rag = "glassbox_rag.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["glassbox_rag*"]

[tool.black]
line-length = 100
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
